+++ info/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2024-03-11
+
+### Added
+
+- Check for dangerous default file extension associations (e.g. '.bat').
+
 ## 2024-03-06
 
 ### Fixed